feat(models): add Customer, PastSubscription, ContactInquiry models and extend existing entities
- Add Customer entity with tenant isolation
- Add PastSubscription for subscription history tracking
- Add ContactInquiry for landing page inquiries
- Add PaginationViewModel for reusable pagination
- Extend Sale with discount and customer link fields
- Extend StockMovement and SaleItem with additional audit fields
- Add IsTenantAdmin / IsSuperAdmin flags to ApplicationUser
- Adjust Plan and TenantSubscription fields for subscription lifecycle

+++ Models/Plan.cs
@@ -17,8 +17,8 @@ namespace ClothInventoryApp.Models
         [MaxLength(500)]
         public string? Description { get; set; }
 
-        public decimal PriceMonthly { get; set; }
-        public decimal? PriceYearly { get; set; }
+        public int PriceMonthly { get; set; }
+        public int? PriceYearly { get; set; }
 
         public int? MaxUsers { get; set; }
         public int? MaxProducts { get; set; }
